+++ app/src/main/java/com/feelsoftware/feelfine/data/model/UserProfile.kt
@@ -3,12 +3,12 @@ package com.feelsoftware.feelfine.data.model
 data class UserProfile(
     val name: String,
     val gender: Gender,
-    val weight: Float,
+    val weight: Int,
     val age: Int,
 ) {
 
     companion object {
-        val EMPTY = UserProfile("", Gender.MALE, 0f, 0)
+        val EMPTY = UserProfile("", Gender.MALE, 0, 0)
     }
 
     enum class Gender {
